feat: 이메일 로그인 기능 구현 및 로그인 페이지 구조 분리 (#51)
- 이메일 로그인 API 연동 및 요청 스키마 적용
- 로그인 실패 케이스별 메시지 처리 구현
- 로그인 성공 시 이전 페이지 우선, 없을 경우 메인으로 리다이렉트
- 로그인 상태 유지 값 요청 바디에 반영
- 네트워크 오류에 대한 예외 처리 추가
- 로그인 페이지 app에서 features/auth/login으로 UI를 조립하는 구조로 분리
- useSearchParams 사용에 따른 Suspense 경계 추가로 빌드 에러 해결
